+++ test/test_helper.rb
@@ -77,6 +77,16 @@ module Test
           user_agent_prefix: user_agent_prefix ? user_agent_prefix : ShopifyAPI::Context.user_agent_prefix
         )
       end
+
+      sig do
+        params(
+          name: String,
+          format: Symbol
+        ).returns(T.nilable(String))
+      end
+      def load_fixture(name, format = :json)
+        File.read(File.dirname(__FILE__) + "/fixtures/#{name}.#{format}")
+      end
     end
   end
 end